% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_express.R
\docType{data}
\name{diff_express}
\alias{diff_express}
\title{diff_express}
\format{
A data frame with 16349 rows and 3 columns.
\describe{
  \item{gene}{gene gene names}
  \item{log2FC}{log2FC log2 fold change}
  \item{pvalue}{p-value}
  ...
}
}
\description{
used as demo data for this package
}
\examples{
data(diff_express)
head(diff_express)
}
